% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{aamutation}
\alias{aamutation}
\alias{aamutation,Variant-method}
\title{aamutation function}
\usage{
aamutation(object)

\S4method{aamutation}{Variant}(object)
}
\arguments{
\item{object}{A Variant Object}
}
\value{
An integer position or '-'
}
\description{
It extracts the AminoAcid position from the variant. If the mutation is a silent mutation or if the info is missing, it returns a "-"
}
\section{Functions}{
\itemize{
\item \code{aamutation(Variant)}: Method for Variant class

}}
