% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{allele.check}
\alias{allele.check}
\title{Validate nucleotide allele
It checks if A given allele sequence is valid according to several criteria}
\usage{
allele.check(
  allele,
  allele.type,
  exp.length = NULL,
  min.length = NULL,
  allow.empty = FALSE
)
}
\arguments{
\item{allele}{A character string representing the allele}

\item{allele.type}{A character string to describe the allele (altAllele, refAllele)}

\item{exp.length}{Optional integer, the expected required length}

\item{min.length}{Optional integer, the minimum required length of the allele}

\item{allow.empty}{Logical, if false, if the allele is empty the function returns an error}
}
\value{
\code{NULL} if the allele is valid, otherwise, a character string with an error
}
\description{
Validate nucleotide allele
It checks if A given allele sequence is valid according to several criteria
}
\keyword{internal}
