% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{altAllele}
\alias{altAllele}
\alias{altAllele,Variant-method}
\title{altAllele accessor}
\usage{
altAllele(object)

\S4method{altAllele}{Variant}(object)
}
\arguments{
\item{object}{a Variant object}
}
\value{
A character string
}
\description{
Get the alternative allele from a Variant Object
}
\section{Methods (by class)}{
\itemize{
\item \code{altAllele(Variant)}: Method for Variant class

}}
\examples{
# altAllele(variant)
}
