% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{altAllele.setter}
\alias{altAllele.setter}
\alias{altAllele<-}
\alias{altAllele<-,Variant-method}
\title{altAllele slot setter}
\usage{
altAllele(object) <- value

\S4method{altAllele}{Variant}(object) <- value
}
\arguments{
\item{object}{A variant object}

\item{value}{A character string representing the new alternative allele}
}
\value{
The new Variant object
}
\description{
Set a new alternative allele
}
\section{Functions}{
\itemize{
\item \code{altAllele(Variant) <- value}: Setter method for the Variant

}}
