% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{clinicalRel}
\alias{clinicalRel}
\alias{clinicalRel,Variant-method}
\title{clinicalRel accessor}
\usage{
clinicalRel(object)

\S4method{clinicalRel}{Variant}(object)
}
\arguments{
\item{object}{a Variant object}
}
\value{
A character string
}
\description{
Get the clinicalRel from a Variant Object
}
\section{Methods (by class)}{
\itemize{
\item \code{clinicalRel(Variant)}: Method for Variant class

}}
\examples{
# clinicalRel(variant)
}
