% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{clinicalRel.setter}
\alias{clinicalRel.setter}
\alias{clinicalRel<-}
\alias{clinicalRel<-,Variant-method}
\title{clinicalRel slot setter}
\usage{
clinicalRel(object) <- value

\S4method{clinicalRel}{Variant}(object) <- value
}
\arguments{
\item{object}{A variant object}

\item{value}{A character string representing the new clinical Relevance}
}
\value{
The new Variant object
}
\description{
Set a new clinical Relevance info
}
\section{Functions}{
\itemize{
\item \code{clinicalRel(Variant) <- value}: Setter method for the Variant

}}
