% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{consequence}
\alias{consequence}
\alias{consequence,Variant-method}
\title{consequence accessor}
\usage{
consequence(object)

\S4method{consequence}{Variant}(object)
}
\arguments{
\item{object}{a Variant object}
}
\value{
A character string
}
\description{
Get the consequence from a Variant Object
}
\section{Methods (by class)}{
\itemize{
\item \code{consequence(Variant)}: Method for Variant class

}}
\examples{
# consequence(variant)
}
