% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{Deletion_validation}
\alias{Deletion_validation}
\title{Validation method for Deletion object}
\arguments{
\item{object}{an \code{Deletion} object}
}
\value{
\code{TRUE} if the object is valid, otherwise a message with the error
}
\description{
It checks if the object is a valid \code{Deletion}, it checks both ref and alt allele
}
\keyword{internal}
