% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{DNV_validation}
\alias{DNV_validation}
\title{Validation method for a DNV object}
\arguments{
\item{object}{a \code{DNV} object}
}
\value{
\code{TRUE} if the object is valid, otherwise a message with the error
}
\description{
It checks if the object is a valid \code{DNV}, it checks both ref and alt allele (exactly 2 nucleotides)
}
\keyword{internal}
