% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{EnsemblID}
\alias{EnsemblID}
\alias{EnsemblID,Variant-method}
\title{EnsemblID accessor}
\usage{
EnsemblID(object)

\S4method{EnsemblID}{Variant}(object)
}
\arguments{
\item{object}{a Variant object}
}
\value{
A character string
}
\description{
Get the EnsemblID from a Variant Object
}
\section{Methods (by class)}{
\itemize{
\item \code{EnsemblID(Variant)}: Method for Variant class

}}
\examples{
# EnsemblID(variant)
}
