% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{EnsemblID.setter}
\alias{EnsemblID.setter}
\alias{EnsemblID<-}
\alias{EnsemblID<-,Variant-method}
\title{EnsemblID slot setter}
\usage{
EnsemblID(object) <- value

\S4method{EnsemblID}{Variant}(object) <- value
}
\arguments{
\item{object}{A variant object}

\item{value}{A character string representing the new EnsemblID}
}
\value{
The new Variant object
}
\description{
Set a new EnsemblID
}
\section{Functions}{
\itemize{
\item \code{EnsemblID(Variant) <- value}: Setter method for the Variant

}}
