% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{geneSymbol}
\alias{geneSymbol}
\alias{geneSymbol,Variant-method}
\title{geneSymbol accessor}
\usage{
geneSymbol(object)

\S4method{geneSymbol}{Variant}(object)
}
\arguments{
\item{object}{a Variant object}
}
\value{
A character string
}
\description{
Get the geneSymbol from a Variant Object
}
\section{Methods (by class)}{
\itemize{
\item \code{geneSymbol(Variant)}: Method for Variant class

}}
\examples{
# geneSymbol(variant)
}
