% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{geneType}
\alias{geneType}
\alias{geneType,Variant-method}
\title{geneType accessor}
\usage{
geneType(object)

\S4method{geneType}{Variant}(object)
}
\arguments{
\item{object}{a Variant object}
}
\value{
A character string
}
\description{
Get the geneType from a Variant Object
}
\section{Methods (by class)}{
\itemize{
\item \code{geneType(Variant)}: Method for Variant class

}}
\examples{
# geneType(variant)
}
