% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{HGVSp}
\alias{HGVSp}
\alias{HGVSp,Variant-method}
\title{HGVSp accessor}
\usage{
HGVSp(object)

\S4method{HGVSp}{Variant}(object)
}
\arguments{
\item{object}{a Variant object}
}
\value{
A character string
}
\description{
Get the HGVSp from a Variant Object
}
\section{Methods (by class)}{
\itemize{
\item \code{HGVSp(Variant)}: Method for Variant class

}}
\examples{
# HGVSp(variant)
}
