% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{HGVSp.setter}
\alias{HGVSp.setter}
\alias{HGVSp<-}
\alias{HGVSp<-,Variant-method}
\title{HGVSp slot setter}
\usage{
HGVSp(object) <- value

\S4method{HGVSp}{Variant}(object) <- value
}
\arguments{
\item{object}{A variant object}

\item{value}{A character string representing the new HGVSp}
}
\value{
The new Variant object
}
\description{
Set a new HGVSp
}
\section{Functions}{
\itemize{
\item \code{HGVSp(Variant) <- value}: Setter method for the Variant

}}
