% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{Insertion_validation}
\alias{Insertion_validation}
\title{Validation method for Insertion object}
\arguments{
\item{object}{an \code{Insertion} object}
}
\value{
\code{TRUE} if the object is valid, otherwise a message with the error
}
\description{
It checks if the object is a valid \code{Insertion}, it checks both ref and alt allele
}
\keyword{internal}
