% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{ONV}
\alias{ONV}
\title{ONV constructor}
\usage{
ONV(
  variantID,
  pos,
  ref,
  alt,
  geneSymb,
  geneType,
  consequence,
  clinicalRel = NA_character_,
  EnsID = NA_character_,
  HGVSp = NA_character_
)
}
\arguments{
\item{variantID}{Character, unique identifier for the variant}

\item{pos}{A Granges object indicating the genomic coordinates}

\item{ref}{Character, reference allele at the specified position}

\item{alt}{Character, alternative allele}

\item{geneSymb}{Character, gene symbol of the mutated gene}

\item{geneType}{Character, type of gene (e.g, protein-coding, lncRNA, etc.)}

\item{consequence}{Character, Functional consequence (e.g. missense, nonsense, silent, frameshift, etc.)}

\item{clinicalRel}{Character, clinical relevance (if available, Clinvar of COSMIC info). Optional}

\item{EnsID}{Character, Ensemble ID of the transcript affected by the variant. Optional}

\item{HGVSp}{HGVS protein level annotation (e.g. p.R381L). Optional}
}
\value{
An object of the class ONV
}
\description{
Creates an object of the class ONV, Oligo-Nucleotide Variant
}
\examples{
#library(GenomicRanges)
#gr <- GenomicRanges::GRanges(seqnames = 'chr1', ranges = IRanges::IRanges(1234, width = 3))
#ONV("rs789", gr, "TAC", "GGA", "EGFR", "protein-coding", "frameshift")
}
