% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{ONV_validation}
\alias{ONV_validation}
\title{Validation method for a ONV object}
\arguments{
\item{object}{a \code{ONV} object}
}
\value{
\code{TRUE} if the object is valid, otherwise a message with the error
}
\description{
It checks if the object is a valid \code{ONV}, it checks both ref and alt allele (at least 3 nucleotides)
}
\keyword{internal}
