% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{position}
\alias{position}
\alias{position,Variant-method}
\title{position accessor}
\usage{
position(object)

\S4method{position}{Variant}(object)
}
\arguments{
\item{object}{a Variant object}
}
\value{
A GRanges object
}
\description{
Get the position from a Variant Object
}
\section{Methods (by class)}{
\itemize{
\item \code{position(Variant)}: Method for Variant class

}}
\examples{
# position(variant)
}
