% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{refAllele}
\alias{refAllele}
\alias{refAllele,Variant-method}
\title{refAllele accessor}
\usage{
refAllele(object)

\S4method{refAllele}{Variant}(object)
}
\arguments{
\item{object}{a Variant object}
}
\value{
A character string
}
\description{
Get the reference allele from a Variant Object
}
\section{Methods (by class)}{
\itemize{
\item \code{refAllele(Variant)}: Method for Variant class

}}
\examples{
# refAllelle(variant)
}
