% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,Variant-method}
\title{Show Variant object
Prints summary of objects of Variant class and their subclasses}
\usage{
\S4method{show}{Variant}(object)
}
\arguments{
\item{object}{an object of the class Variant or any class that inherits from it}
}
\value{
None
It prints a formatted summary of the object to the console
}
\description{
Show Variant object
Prints summary of objects of Variant class and their subclasses
}
