% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\docType{class}
\name{SNV-class}
\alias{SNV-class}
\title{SNV class}
\description{
Class for Single Nucleotide Variants
This class inherits from the \code{\link{Variant-class}} virtual class
}
\examples{
library(GenomicRanges)
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, 100))
snv <- new("SNV", variantID="rs1", position=gr, refAllele="A", altAllele="T",
           geneSymbol="MYC", geneType="protein_coding", consequence="missense",
           clinicalRel=NA_character_, EnsemblID=NA_character_, HGVSp=NA_character_)
snv
}
