% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{SNV_validation}
\alias{SNV_validation}
\title{Validation method for a SNV object}
\arguments{
\item{object}{a \code{SNV} object}
}
\value{
\code{TRUE} if the object is valid, otherwise a message with the error
}
\description{
It checks if the object is a valid \code{SNV}, it checks both ref and alt allele (exactly 1 nucleotide)
}
\keyword{internal}
