% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\docType{class}
\name{Variant-class}
\alias{Variant-class}
\title{Variant Virtual Class}
\description{
virtual base class for all mutation types
}
\section{Slots}{

\describe{
\item{\code{variantID}}{Character. Unique ID for the variant}

\item{\code{position}}{GRanges. Genomics Coordinates}

\item{\code{refAllele}}{Character. Reference Allele}

\item{\code{altAllele}}{Character. Alternative Allele}

\item{\code{geneSymbol}}{Character. Symbol of the affected gene}

\item{\code{geneType}}{Character. Type of the mutated gene}

\item{\code{consequence}}{Character. Type of consequence - effect}

\item{\code{clinicalRel}}{Character. Clinical Relevance information}

\item{\code{EnsemblID}}{Character. Ensembl transcript ID}

\item{\code{HGVSp}}{Character. Amino Acid change in HGVS format}
}}

