% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_.R
\name{variantID}
\alias{variantID}
\alias{variantID,Variant-method}
\title{variantID accessor}
\usage{
variantID(object)

\S4method{variantID}{Variant}(object)
}
\arguments{
\item{object}{a Variant object}
}
\value{
A character string
}
\description{
Get the variantID from a Variant Object
}
\section{Methods (by class)}{
\itemize{
\item \code{variantID(Variant)}: Method for Variant class

}}
\examples{
# variantID(variant)
}
